+++ clients/admin-ui/src/features/data-discovery-and-detection/hooks/useDiscoveryRoutes.tsx
@@ -12,16 +12,16 @@ const useDiscoveryRoutes = () => {
 
   const navigateToDetectionResults = ({
     resourceUrn,
-    showFullSchema = false,
+    filterTabIndex,
   }: {
     resourceUrn: string;
-    showFullSchema?: boolean;
+    filterTabIndex?: number;
   }) => {
     router.push({
       pathname: DATA_DETECTION_ROUTE_DETAIL,
       query: {
         resourceUrn,
-        showFullSchema,
+        filterTabIndex,
       },
     });
   };